improve package gstr
